minirt restored

+++ minirt/includes/camera.h
@@ -4,6 +4,6 @@
 # include "structs.h"
 
 void		print_camera(t_camera cam);
-t_camera	set_camera(double hsize, double vsize, double fov);
+t_camera	set_camera(double hsize, double vsize, float fov);
 t_ray		ray_for_pixel(t_camera cam, double px, double py);
 #endif
